# mes-root/docs/mes-auto_swagger.yaml
openapi: 3.0.0
...
servers:
  - description: DEV
    url: https://localhost:9998
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/hengyi/mes-auto/1.0.0
info:
  description: auto api
  version: "1.0.0"
  title: Mes Auto
  contact:
    email: ixtf1984@gmail.com
...
paths:
...
  /api/silkBarcodes:
    post:
      tags:
        - api
      summary: 丝锭条码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SilkBarcodeGenerateCommand'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - api
      summary: 丝锭条码查询
      parameters:
        - in: query
          name: first
          schema:
            type: integer
            minimum: 0
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 10
            maximum: 100
        - in: query
          name: codeDate
          schema:
            type: string
            format: date
        - in: query
          name: lineMachineId
          schema:
            type: string
            description: 机台
        - in: query
          name: doffingNum
          schema:
            type: string
            description: 落次
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  silkBarcodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/SilkBarcode'
        '400':
          $ref: '#/components/responses/JException'
...
  /api/batchSilkBarcodes:
    post:
      tags:
        - api
      summary: 批量丝锭条码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SilkBarcodeGenerateCommand'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
...
  /api/packageBoxs/barcodes/{code}:
    get:
      tags:
        - api
      summary: 唛头信息
      parameters:
        - in: path
          name: code
          required: true
          description: 唛头编号
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBox'
        '400':
          $ref: '#/components/responses/JException'
...
  /api/reports/monthPackageBoxReport:
    get:
      tags:
        - api
      summary: 月打包
      parameters:
        - in: path
          name: workshopId
          required: true
          schema:
            type: string
        - in: path
          name: year
          required: true
          schema:
            type: integer
        - in: path
          name: month
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBoxReport'
        '400':
          $ref: '#/components/responses/JException'
  /api/reports/dailyPackageBoxReport:
    get:
      tags:
        - api
      summary: 日打包
      parameters:
        - in: path
          name: workshopId
          required: true
          schema:
            type: string
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  packageBoxes:
                    type: array
                    items:
                      $ref: '#/components/schemas/PackageBox'
        '400':
          $ref: '#/components/responses/JException'
  /api/reports/dailyDoffingReport:
    get:
      tags:
        - api
      summary: 日下机
      parameters:
        - in: path
          name: workshopId
          required: true
          schema:
            type: string
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoffingReport'
        '400':
          $ref: '#/components/responses/JException'
  /api/reports/dailySilkExceptionReport:
    get:
      tags:
        - api
      summary: 日异常
      parameters:
        - in: path
          name: workshopId
          required: true
          schema:
            type: string
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SilkExceptionReport'
        '400':
          $ref: '#/components/responses/JException'
  /api/reports/monthSilkExceptionReport:
    get:
      tags:
        - api
      summary: 月异常
      parameters:
        - in: path
          name: workshopId
          required: true
          schema:
            type: string
        - in: path
          name: year
          required: true
          schema:
            type: integer
        - in: path
          name: month
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  packageBoxes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Silk'
        '400':
          $ref: '#/components/responses/JException'
components:
  securitySchemes:
    # arbitrary name for the security scheme
    bearerAuth:
      type: http
      scheme: bearer
      # optional, arbitrary value for documentation purposes
      bearerFormat: JWT
  responses:
...
    JException:
      description: |
        非200的异常一般都会处理过，没处理的都是http status默认

        如：401没认证，403没权限，404找不到，等等
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JException'
  schemas:
    JException:
      type: object
      properties:
        errorCode:
          type: string
          example: E00000
        errorMessage:
          type: string
...
    BasicPermission:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
...
    EntityDTO:
      type: object
      required:
        - id
      properties:
        id:
          type: string
...
    SilkBarcodeGenerateCommand:
      type: object
      required:
        - lineMachine
        - codeDate
        - doffingNum
      properties:
        lineMachine:
          $ref: '#/components/schemas/EntityDTO'
        codeDate:
          type: string
          format: date
        doffingNum:
          type: string
          description: 落次
    SilkBarcode:
      type: object
      properties:
        batch:
          $ref: '#/components/schemas/Batch'
        lineMachine:
          $ref: '#/components/schemas/LineMachine'
        codeDate:
          type: string
          format: date
        doffingNum:
          type: string
          description: 落次
        silks:
          type: array
          items:
            $ref: '#/components/schemas/Silk'
...
    Operator:
      allOf:
        - $ref: '#/components/schemas/BasicPermission'
        - type: object
          properties:
            id:
              type: string
              readOnly: true
            hrId:
              type: string
              readOnly: true
            oaId:
              type: string
              readOnly: true
            name:
              type: string
              readOnly: true
            admin:
              type: boolean
            groups:
              type: array
              items:
                $ref: '#/components/schemas/OperatorGroup'
...
    OperatorGroup:
      allOf:
        - $ref: '#/components/schemas/BasicPermission'
        - type: object
          properties:
            id:
              type: string
            name:
              type: string
    Corporation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    Workshop:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        note:
          type: string
        corporation:
          $ref: '#/components/schemas/Corporation'
...
    Line:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        autoDoffing:
          type: boolean
        workshop:
          $ref: '#/components/schemas/Workshop'
...
    LineMachine:
      type: object
      properties:
        id:
          type: string
        line:
          $ref: '#/components/schemas/Line'
        item:
          type: integer
        spindleNum:
          type: integer
          description: 锭数
        spindleSeq:
          description: 落筒的锭位顺序
          type: array
          items:
            type: integer
...
    Silk:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        doffingNum:
          type: string
          description: 落次
        doffingOperator:
          $ref: '#/components/schemas/Operator'
...
    Product:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
...
    Batch:
      type: object
      description: 批号
      properties:
        id:
          type: string
        batchNo:
          type: string
        silkWeight:
          type: number
        centralValue:
          type: number
        holeNum:
          type: number
        spec:
          type: string
          description: 规格
        tubeColor:
          type: string
          description: 规格
        note:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        product:
          $ref: '#/components/schemas/Product'
...
    PackageClass:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        packageCode:
          type: string
        sortBy:
          type: string
...
    PackageBox:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
          description: 唛头条码
        palletCode:
          type: string
          description: 久鼎塑托
        type:
          type: string
          enum: [AUTO, MANUAL,FOREIGN]
          description: 打包类型，自动打包，人工打包，外贸打包
        grossWeight:
          type: number
          description: 毛重
        netWeight:
          type: number
          description: 净重
        batch:
          $ref: '#/components/schemas/Batch'
        printClass:
          description: 打印班次
          $ref: '#/components/schemas/PackageClass'
        printDate:
          type: number
          description: 打印日期
        budatClass:
          description: 入库班次
          $ref: '#/components/schemas/PackageClass'
        budat:
          type: number
          description: 入库日期
        silkCount:
          type: integer
          description: 丝锭颗数
        silks:
          type: array
          items:
            $ref: '#/components/schemas/Silk'
        createDateTime:
          type: number
          description: 打包时间
...
    PackageBoxReport:
      type: object
      properties:
        packageBoxes:
          type: array
          items:
            $ref: '#/components/schemas/PackageBox'
    SilkExceptionReport:
      type: object
      properties:
        packageBoxes:
          type: array
          items:
            $ref: '#/components/schemas/Silk'
    DoffingReport:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              batch:
                $ref: '#/components/schemas/Batch'
              silks:
                type: array
                items:
                  $ref: '#/components/schemas/Silk'
...
security:
  - bearerAuth: []